feat: 채팅방 생성 기회 조작 코드 추가

+++ src/main/java/com/goorm/team9/icontact/domain/client/service/ClientService.java
@@ -40,6 +40,7 @@ public class ClientService {
         clientEntity.setStatus(status);
         clientEntity.setIntroduction(introduction);
         clientEntity.setLink(link);
+        clientEntity.setChatOpportunity(5);
 
         if (profileImage != null && !profileImage.isEmpty()) {
             String imagePath = imageFileStorageService.storeFile(profileImage);
@@ -112,15 +113,40 @@ public class ClientService {
             existingClient.setProfileImage(imagePath);
         }
 
-        if (profileImage == null) {
-            if (!imageFileStorageService.isDefaultImage(currentImage)) {
-                imageFileStorageService.deleteFile(currentImage); // 기존 이미지 삭제
-            }
-            existingClient.setProfileImage("/profile-images/default_profile_image.jpg"); // 기본 이미지로 변경
+        if (profileImage == null && !imageFileStorageService.isDefaultImage(currentImage)) {
+            imageFileStorageService.deleteFile(currentImage);
+            existingClient.setProfileImage("/profile-images/default_profile_image.jpg");
         }
 
         ClientEntity updatedClient = clientRepository.save(existingClient);
         return clientConverter.toResponseDTO(updatedClient);
     }
+
+    @Transactional
+    public void reduceChatOpportunity(Long clientId) {
+        ClientEntity clientEntity = clientRepository.findById(clientId)
+                .orElseThrow(() -> new CustomException(ClientErrorCode.CLIENT_NOT_FOUND));
+
+        if (clientEntity.getChatOpportunity() <= 0) {
+            throw new CustomException(ClientErrorCode.NO_CHAT_OPPORTUNITY);
+        }
+
+        clientEntity.setChatOpportunity(clientEntity.getChatOpportunity() - 1);
+        clientRepository.save(clientEntity);
+    }
+
+    @Transactional
+    public void increaseChatOpportunity(Long clientId) {
+        ClientEntity clientEntity = clientRepository.findById(clientId)
+                .orElseThrow(() -> new CustomException(ClientErrorCode.CLIENT_NOT_FOUND));
+
+        if (clientEntity.getChatOpportunity() >= 5) {
+            throw new CustomException(ClientErrorCode.CHAT_OPPORTUNITY_FULL);
+        }
+
+        clientEntity.setChatOpportunity(clientEntity.getChatOpportunity() + 1);
+        clientRepository.save(clientEntity);
+    }
+
 
 }
